R1: move rule_rewrite.py -> ir/hir/rule_rewrite.py
HIR-level passes (PassInfo level=RULE_REWRITE, source/target_dialect=HIR)
sitting at the top level was inconsistent with the rest of the HIR
pass framework in ir/hir/{index,plan,semi_naive,split,stratify}.py.
Move it next to its peers.

No behavior change. Updates 3 import sites:
  src/srdatalog/ir/hir/__init__.py  (default_pipeline)
  tests/test_rule_rewrite.py
  tests/test_semi_join.py
