+++ tests/test_semi_join.py
@@ -10,7 +10,7 @@ from srdatalog.ir.hir import compile_to_hir, compile_to_mir
 from srdatalog.ir.hir.emit import hir_to_obj
 from srdatalog.ir.mir.emit import print_mir_sexpr
 from srdatalog.provenance import ProvenanceKind
-from srdatalog.rule_rewrite import _is_semi_join_candidate, optimize_semi_joins
+from srdatalog.ir.hir.rule_rewrite import _is_semi_join_candidate, optimize_semi_joins
 
 FIXTURES = Path(__file__).resolve().parent / "fixtures"
